Addresse PR's comments: remove code re-ordering and rename SemanticMetricRegistry.getDistribution
