+++ core/src/main/java/com/spotify/metrics/core/SemanticMetricBuilder.java
@@ -27,7 +27,6 @@ import com.codahale.metrics.Histogram;
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.Metric;
 import com.codahale.metrics.Timer;
-import com.spotify.metrics.core.codahale.metrics.ext.Distribution;
 
 /**
  * A quick and easy way of capturing the notion of default metrics.
