+++ core/src/main/java/com/spotify/metrics/core/SemanticMetricRegistryListener.java
@@ -35,7 +35,6 @@ import com.codahale.metrics.Gauge;
 import com.codahale.metrics.Histogram;
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.Timer;
-import com.spotify.metrics.core.codahale.metrics.ext.Distribution;
 
 import java.util.EventListener;
 
@@ -43,6 +42,69 @@ import java.util.EventListener;
  * Listeners for events from the registry.  Listeners must be thread-safe.
  */
 public interface SemanticMetricRegistryListener extends EventListener {
+    /**
+     * A no-op implementation of {@link SemanticMetricRegistryListener}.
+     */
+    abstract class Base implements SemanticMetricRegistryListener {
+        @Override
+        public void onGaugeAdded(MetricId name, Gauge<?> gauge) {
+        }
+
+        @Override
+        public void onGaugeRemoved(MetricId name) {
+        }
+
+        @Override
+        public void onCounterAdded(MetricId name, Counter counter) {
+        }
+
+        @Override
+        public void onCounterRemoved(MetricId name) {
+        }
+
+        @Override
+        public void onHistogramAdded(MetricId name, Histogram histogram) {
+        }
+
+        @Override
+        public void onHistogramRemoved(MetricId name) {
+        }
+
+        @Override
+        public void onMeterAdded(MetricId name, Meter meter) {
+        }
+
+        @Override
+        public void onMeterRemoved(MetricId name) {
+        }
+
+        @Override
+        public void onTimerAdded(MetricId name, Timer timer) {
+        }
+
+        @Override
+        public void onTimerRemoved(MetricId name) {
+        }
+
+        @Override
+        public void onDerivingMeterAdded(MetricId name, DerivingMeter derivingMeter) {
+        }
+
+        @Override
+        public void onDerivingMeterRemoved(MetricId name) {
+        }
+
+        @Override
+        public void onDistributionAdded(MetricId name, Distribution distribution) {
+
+        }
+
+        @Override
+        public  void onDistributionRemoved(MetricId name) {
+
+        }
+    }
+
     /**
      * Called when a {@link Gauge} is added to the registry.
      *
@@ -121,7 +183,7 @@ public interface SemanticMetricRegistryListener extends EventListener {
     /**
      * Called when a {@link DerivingMeter} is added to the registry.
      *
-     * @param name          the meter's name
+     * @param name  the meter's name
      * @param derivingMeter the meter
      */
     void onDerivingMeterAdded(MetricId name, DerivingMeter derivingMeter);
@@ -155,68 +217,4 @@ public interface SemanticMetricRegistryListener extends EventListener {
     public default void onDistributionRemoved(MetricId name) {
 
     }
-
-    /**
-     * A no-op implementation of {@link SemanticMetricRegistryListener}.
-     */
-    abstract class Base implements SemanticMetricRegistryListener {
-        @Override
-        public void onGaugeAdded(MetricId name, Gauge<?> gauge) {
-        }
-
-        @Override
-        public void onGaugeRemoved(MetricId name) {
-        }
-
-        @Override
-        public void onCounterAdded(MetricId name, Counter counter) {
-        }
-
-        @Override
-        public void onCounterRemoved(MetricId name) {
-        }
-
-        @Override
-        public void onHistogramAdded(MetricId name, Histogram histogram) {
-        }
-
-        @Override
-        public void onHistogramRemoved(MetricId name) {
-        }
-
-        @Override
-        public void onMeterAdded(MetricId name, Meter meter) {
-        }
-
-        @Override
-        public void onMeterRemoved(MetricId name) {
-        }
-
-        @Override
-        public void onTimerAdded(MetricId name, Timer timer) {
-        }
-
-        @Override
-        public void onTimerRemoved(MetricId name) {
-        }
-
-        @Override
-        public void onDerivingMeterAdded(MetricId name, DerivingMeter derivingMeter) {
-        }
-
-        @Override
-        public void onDerivingMeterRemoved(MetricId name) {
-        }
-
-        @Override
-        public void onDistributionAdded(MetricId name, Distribution distribution) {
-
-        }
-
-        @Override
-        public  void onDistributionRemoved(MetricId name) {
-
-        }
-    }
-
 }
